% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_MCMCloop_withDENVimmunity_adapt.R
\name{AdaptMCMCloop_withDENVimmunity}
\alias{AdaptMCMCloop_withDENVimmunity}
\title{MCMC function for fitting Zika model using MH algorithm}
\usage{
AdaptMCMCloop_withDENVimmunity(sample.start.point = T,
  startdate = as.Date("2013-10-28"), adapt.shape.start = MCMC.runs)
}
\arguments{
\item{startdate}{Original start date in the data series}

\item{sammple.start.point}{Whether or not to sample start time}
}
\description{
MCMC function for fitting Zika model using MH algorithm
}
