% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculateR0.R
\name{calculate_r0}
\alias{calculate_r0}
\title{calculate_r0}
\usage{
calculate_r0(th_in, sus_c = 1, sus_a = 1, sm_c = 1, sm_a = 1,
  b_vary = 1)
}
\arguments{
\item{th_in}{Binary indicator variable if model is age structured (1) or not (0) between children and adults. Defaults to NULL in which case no sampling happens}

\item{sus_c}{Suceptible child human hosts}

\item{sus_a}{Suceptible adult human hosts}

\item{sm_c}{Suceptible child moisquitoes}

\item{sm_a}{Suceptible adult moisquitoes}

\item{b_vary}{Defaults to 1}
}
\description{
calculate_r0
}
