% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ComputePrior.R
\name{ComputePrior}
\alias{ComputePrior}
\title{Compute prior}
\usage{
ComputePrior(iiH, thetaAlltab, thetaAllstar)
}
\arguments{
\item{iiH}{Location}

\item{thetaAlltab}{Current values of theta}

\item{thetaAllstar}{Proposed values of theta}
}
\description{
Compute prior
}
