% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ComputePrior_den2.R
\name{ComputePrior_D2}
\alias{ComputePrior_D2}
\title{Compute prior}
\usage{
ComputePrior_D2(iiH, thetaAlltab, thetaAllstar)
}
\arguments{
\item{iiH}{Location}

\item{thetaAlltab}{Current values of theta}

\item{thetaAllstar}{Proposed values of theta}
}
\description{
Compute prior
}
