% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ComputePrior_special.R
\name{ComputePrior_special}
\alias{ComputePrior_special}
\title{Compute prior}
\usage{
ComputePrior_special(iiH, thetaAlltab, thetaAllstar, parameter_est_file)
}
\arguments{
\item{iiH}{Location}

\item{thetaAlltab}{Current values of theta}

\item{thetaAllstar}{Proposed values of theta}

\item{parameters_est_file}{Name of file with list of parameters to estimate}
}
\description{
Compute prior
}
