% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_seasonal_decline.R
\name{control_f}
\alias{control_f}
\title{Flexible decline function transmission rate function}
\usage{
control_f(time, mask = 0, base = 1, grad = 0.1, mid = 0,
  mid2 = 1, width = 30)
}
\arguments{
\item{time}{t in time.vals of calc}

\item{mask}{Strength of reduction in transmission rate. Defaults to 0 (i.e. no effect)}

\item{base}{Height of starting point of function. Defaults to 1}

\item{grad}{Strength of decline in transmission once begun. Defaults to 0.5}

\item{mid}{Point at which decline can begin. Defaults to 0}

\item{date0}{reference index value for start of transmission. Defaults to 0}
}
\description{
Sigmoid function to allow drop in transmission rate
}
\examples{

}
\keyword{Decline}
\keyword{Sigmoid}
