% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_seasonal_decline.R
\name{CrossImmune}
\alias{CrossImmune}
\title{Cross immunity}
\usage{
CrossImmune(time, date0 = 0, mask = 1, base = 0.2, grad = 2,
  mid = 0.3, wane = 1)
}
\arguments{
\item{immune}{}

\item{pop}{}
}
\description{
Given a level of dengue infection, susceptible Zika hosts are depressed
}
