% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_seasonal_decline.R
\name{death_f_log}
\alias{death_f_log}
\title{Flexible decline function transmission rate function}
\usage{
death_f_log(time, date0 = 0, mask = 1, base = 0.5, grad = 1e+06,
  mid = 1)
}
\arguments{
\item{time}{t in time.vals of calc}

\item{date0}{reference index value for start of transmission. Defaults to 0}

\item{mask}{Strength of reduction in transmission rate. Defaults to 0 (i.e. no effect)}

\item{base}{Height of starting point of function. Defaults to 1}

\item{grad}{Strength of decline in transmission once begun. Defaults to 0.5}

\item{mid}{Point at which decline can begin. Defaults to 0}

\item{date.interventions}{Vector of dates when vector control interventions took place}
}
\description{
Sigmoid function to allow drop in transmission rate
}
\keyword{Decline}
\keyword{Sigmoid}
