% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_det_modelR_DENVimm.R
\name{Deterministic_modelR_final_DENVimmmunity}
\alias{Deterministic_modelR_final_DENVimmmunity}
\title{Deterministic SEIR-SEI function with no age structure}
\usage{
Deterministic_modelR_final_DENVimmmunity(agestructure = NULL, theta,
  theta_init, locationI, seroposdates, episeason, include.count = T)
}
\arguments{
\item{agestructure}{Binary indicator variable if model is age structured (1) or not (0) between children and adults. Defaults to NULL in which case no sampling happens}

\item{theta}{Vector of parameters for model}

\item{theta_init}{Vector of initial conditions for model}

\item{locationI}{Name of location of data}

\item{seroposdates}{Vector with dates of seroprevalence surveys}

\item{episeason}{Vector with start and end point of epidemic case data}

\item{include.count}{True or False - whether to include count data in likelihood. Defaults to True}
}
\description{
This function solves an SEIR-SEI vector borne disease model.
}
\examples{
Deterministic_modelR(c(thetatab[1,],thetaAlltab[1,iiH,]), theta_initAlltab[1,iiH,],locationI=locationtab[iiH],seroposdates= c(as.Date("2013-11-25"), as.Date("2015-11-02")), episeason=c(as.Date("2016-02-22"), as.Date("2016-06-06")))
}
\keyword{deterministic}
