% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_seasonal_decline.R
\name{extinct}
\alias{extinct}
\title{Reporting cases function}
\usage{
extinct(x, x0)
}
\arguments{
\item{x}{number to test}

\item{x0}{value to be greater than}
}
\description{
Given a number of infected hosts, this function estimates the resulting number of cases
}
