% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_adjust_start_time.R
\name{fn_adjust_start_time}
\alias{fn_adjust_start_time}
\title{Function to adjusst start time and data when sampling start times}
\usage{
fn_adjust_start_time(sample.start.point = F, t0)
}
\arguments{
\item{sample.start.point}{True/False whether to sample start time. Defaults to F}

\item{t0}{New sampled value of t0}
}
\description{
Funciton that reloads data and fills in zeroes for time series and date series if sampling start time
}
