% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ImmunityFit.R, R/fn_WeatherFit.R
\name{likelihood_fn}
\alias{likelihood_fn}
\title{WeatherFit likelihood fn}
\usage{
likelihood_fn(amp, mid, time.vals = time.vals, beta = beta, data = data)

likelihood_fn(amp, mid, time.vals = time.vals, beta = beta, data = data)
}
\arguments{
\item{amp}{amplitude of seasonal wave. Defaults to 0}

\item{mid}{midpoint of seasonal wave. Defaults to -0.048 to fix at Feb}

\item{time.vals}{Time series over which to fit sin curve to data}

\item{beta}{Average value from data to act as midpoint of sincurve}

\item{data}{Data values}

\item{amp}{amplitude of seasonal wave. Defaults to 0}

\item{mid}{midpoint of seasonal wave. Defaults to -0.048 to fix at Feb}

\item{time.vals}{Time series over which to fit sin curve to data}

\item{beta}{Average value from data to act as midpoint of sincurve}

\item{data}{Data values}
}
\description{
Likelihood for seasonality parameters fitting to temperature data

Likelihood for seasonality parameters fitting to temperature data
}
\keyword{Seasonality}
