% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_LoadMultistartData.R
\name{load.data.multistart}
\alias{load.data.multistart}
\title{Load data for model fitting}
\usage{
load.data.multistart(add.nulls = 0, startdate, Virus, cases.file.name,
  serology.file.name, init.values.file.name)
}
\arguments{
\item{add.nulls}{Number of time series values to add to end (default is 0)}

\item{startdate}{Start month of time series}

\item{Virus}{Name of virus for serology results. e.g. 'Zika'}

\item{cases.file.name}{Name of csv file with surveillance data}

\item{serology.file.name}{Name of csv file with serological data}

\item{init.values.file.name}{Name of csv file with initial parameter values}
}
\description{
This function loads surveillance cases, serology results and initial parameter values for vector-borne model fitting
}
\keyword{import}
