% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_LoadData.R
\name{load.data}
\alias{load.data}
\title{Load data for model fitting}
\usage{
load.data(agestructure = NULL, Virus, cases.file.name, serology.file.name,
  init.values.file.name)
}
\arguments{
\item{agestructure}{Binary indicator variable if model is age structured (1) or not (0) between children and adults. Defaults to NULL in which case no sampling happens}

\item{Virus}{Name of virus for serology results. e.g. 'Zika'}

\item{cases.file.name}{Name of csv file with surveillance data}

\item{serology.file.name}{Name of csv file with serological data}

\item{init.values.file.name}{Name of csv file with initial parameter values}

\item{month.start}{Start month of time series}
}
\description{
This function loads surveillance cases, serology results and initial parameter values for vector-borne model fitting
}
