% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_LoadPosteriors.R
\name{load.posteriors}
\alias{load.posteriors}
\title{load.posteriors}
\usage{
load.posteriors(agestructure, iiH, mcmc.burn)
}
\arguments{
\item{agestructure}{Binary indicator variable if model is age structured (1) or not (0) between children and adults. Defaults to NULL in which case no sampling happens}

\item{iiH}{Locationtab index}

\item{mcmc.burn}{Between 0 and 1. Proportion of iterations to discard.}
}
\description{
Load and combine posterior chains following MCMC run
}
