% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_MCMCloop.R
\name{MCMCloop}
\alias{MCMCloop}
\title{MCMC function for a single iteration using MH algorithm}
\usage{
MCMCloop(agestructure, include.count = T)
}
\arguments{
\item{agestructure}{Binary indicator variable if model is age structured (1) or not (0) between children and adults. Defaults to NULL in which case no sampling happens}

\item{include.count}{True or False - whether to include count data in likelihood. Defaults to True}
}
\description{
MCMC function for a single iteration using MH algorithm
}
