% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_MCMCloop_withDENVimmunity.R
\name{MCMCloop_withDENVimmunity}
\alias{MCMCloop_withDENVimmunity}
\title{MCMC function for fitting Zika model using MH algorithm}
\usage{
MCMCloop_withDENVimmunity(sample.start.point = T,
  startdate = as.Date("2013-10-28"))
}
\arguments{
\item{sample.start.point}{Whether or not to sample start time}

\item{startdate}{Original start date in the data series}
}
\description{
MCMC function for fitting Zika model using MH algorithm
}
