% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_PlotMainfigure.R
\name{plot_figure1}
\alias{plot_figure1}
\title{Plot_figure1}
\usage{
plot_figure1(agestructure, btsp, location = 1, seroposdates, nplots = 2)
}
\arguments{
\item{agestructure}{Binary indicator variable if model is age structured (1) or not (0) between children and adults. Defaults to NULL in which case no sampling happens}

\item{btsp}{Number of iterations from MCMC loop to include}

\item{location}{Number of locations to loop over. Defaults to 1}

\item{seroposdates}{Should be same as in MCMC loop}

\item{nplots}{Number of plots - 2 or 3? Defaults to 2}
}
\description{
Plot_figure1
}
