% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_PlotPosteriors.R
\name{plot_Sposteriors}
\alias{plot_Sposteriors}
\title{Plot_Sposteriorss}
\usage{
plot_Sposteriors(locnnLoop = 1, max_season = 2, plot.split = 6,
  param.names, param.labels)
}
\arguments{
\item{max_season}{Defaults to 2}

\item{plot.split}{Max number of params in one correlation plots}

\item{param.names}{vector of names of parameters to plot}

\item{param.labels}{vector of labels to put on corr plot}

\item{locnnloop}{number of locations to loop over. Defaults to 1}
}
\description{
Plots of prior and posterior densities, ESS and correlation between parameter posteriors
}
