% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_seasonal_decline.R
\name{ReportC}
\alias{ReportC}
\title{Reporting cases function}
\usage{
ReportC(cases, rep, repvol)
}
\arguments{
\item{cases}{List of cases from model}

\item{rep}{Reporting rate used to generate those cases}

\item{repvol}{Overdispersion parameter used in Likelihood to estimate those cases}
}
\description{
Given a number of infected hosts, this function estimates the resulting number of cases
}
