% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ResultsSetUp_Multiloc.R
\name{results.set.up.Multiloc}
\alias{results.set.up.Multiloc}
\title{Setup results vectors and dataframes and load initial values. Totally susceptible initially}
\usage{
results.set.up.Multiloc(agestructure, iiH, parameter_est_file)
}
\arguments{
\item{agestructure}{Binary indicator variable if model is age structured (1) or not (0) between children and adults. Defaults to NULL in which case no sampling happens}

\item{iiH}{PLace in locationtab of location being processed}

\item{parameter_est_file}{Name of csv file with parameters to be estimated included}
}
\description{
Setup results vectors and dataframes and load initial values. Totally susceptible initially
}
