% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ResultsSetUp.R
\name{results_set_up}
\alias{results_set_up}
\title{Setup results vectors and dataframes and load initial values. Totally susceptible initially}
\usage{
results_set_up(iiH, parameter_est_file)
}
\arguments{
\item{iiH}{PLace in locationtab of location being processed}

\item{parameter_est_file}{Name of csv file with parameters to be estimated included}
}
\description{
Setup results vectors and dataframes and load initial values. Totally susceptible initially
}
