% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_DENV_MCMC.R
\name{Run_simulation}
\alias{Run_simulation}
\title{Simulate function for SIR}
\usage{
Run_simulation(dt, theta, theta_init, time.vals)
}
\arguments{
\item{dt}{time step}

\item{theta}{vector of parameter values. Must include npop1, beta, gamma and rep}

\item{theta_init}{vector of initial valeus for S I R and C}

\item{time.vals}{Time series to estimate over}
}
\description{
Simulate function for SIR
}
