% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_SampleThetaPartial.R
\name{SampleThetaPartial}
\alias{SampleThetaPartial}
\title{Sample the vector of parameters from a multivariate normal distribution}
\usage{
SampleThetaPartial(theta_in, theta_init_in, covartheta, covartheta_init,
  global = NULL)
}
\arguments{
\item{theta_in}{Vector of previous values of parameters}

\item{theta_init_in}{Vector of previous values of initial conditions}

\item{covartheta}{Covariance matrix for parameters. Must be same width as length of theta_in}

\item{covartheta_init}{Covariance matrix for initial conditions. Must be same width as length of theta_init_in}

\item{global}{Binary indicator variable if sampling global (1) or local (0) parameters. Defaults to NULL in which case no sampling happens}
}
\description{
Sample a new vector of parameters (either global or local) and initial conditions using a multivariate normal dist
}
