% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_seasonal_decline.R
\name{seasonal_f}
\alias{seasonal_f}
\title{Seasonal transmission rate function}
\usage{
seasonal_f(time, date0 = 0, amp = 0, mid = pi * (3/4))
}
\arguments{
\item{time}{t in time.vals of calc}

\item{date0}{reference index value for start of transmission. Defaults to 0}

\item{amp}{amplitude of seasonal wave. Defaults to 0}

\item{mid}{midpoint of seasonal wave. Defaults to -0.048 to fix at Feb}
}
\description{
Calculates the seasonal tranmsission rate
}
\examples{
seasonal_f(date0=0, amp=theta[["beta_v_amp"]], mid=theta[["beta_v_mid"]])
}
\keyword{Seasonality}
