% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sim_det_mosqPop.R
\name{simulate_deterministic_mosqPop}
\alias{simulate_deterministic_mosqPop}
\title{ODEs for SEIR-SEI with noage structure}
\usage{
simulate_deterministic_mosqPop(theta, init.state, time.vals.sim)
}
\arguments{
\item{theta}{Vector of parameters for model}

\item{init.state}{List of initial values}

\item{time.vals.sim}{List of time values to simulate over}
}
\description{
This function solves an SEIR-SEI vector borne disease model.
}
