% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_simulate_deterministic_noage_denvimmunity.R
\name{simulate_deterministic_noage_DENVimm}
\alias{simulate_deterministic_noage_DENVimm}
\title{ODEs for SEIR-SEI with cross immunity from denv infection}
\usage{
simulate_deterministic_noage_DENVimm(theta, init.state, time.vals.sim)
}
\arguments{
\item{theta}{Vector of parameters for model}

\item{init.state}{List of initial values}

\item{time.vals.sim}{List of time values to simulate over}
}
\description{
This function solves an SEIR-SEI vector borne disease model.
}
