% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ggplottheme_zikafiji.R
\name{theme_zika_fiji}
\alias{theme_zika_fiji}
\title{Background theme for ggplots for Zika paper}
\usage{
theme_zika_fiji(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{default = 11}

\item{base_family}{default = ""}
}
\description{
Background ggplot theme for all plots for Zika paper
}
