% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_WeatherFit.R
\name{weather.fit}
\alias{weather.fit}
\title{WeatherFit}
\usage{
weather.fit(data, iter, tuning)
}
\arguments{
\item{iter}{number of iterations to fit over}

\item{tuning}{Tuning param for proposal distributions}

\item{time.length}{timeseries to fit over}

\item{timeseries}{timeseries to fit over}

\item{data.series}{data series to fit to}
}
\description{
Fit seasonality parameters to temperature data
}
\keyword{Seasonality}
