% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ComputePrior.R
\name{ComputePrior}
\alias{ComputePrior}
\title{Compute prior}
\usage{
ComputePrior(iiH, thetaAlltab, thetaAllstar, covartheta)
}
\arguments{
\item{iiH}{Numeric indicator of location given by location vector "locationtab[]"}

\item{thetaAlltab}{Current values of parameter set theta}

\item{thetaAllstar}{Proposed values of parameter set theta}
}
\description{
Compute prior
}
