% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_seasonal_decline.R
\name{decline_f}
\alias{decline_f}
\title{Flexible decline function transmission rate function}
\usage{
decline_f(time, mid, width, base)
}
\arguments{
\item{time}{t in time.vals of calc}

\item{mid}{Point at which decline can begin. Defaults to 0}

\item{base}{Height of starting point of function. Defaults to 1}

\item{date0}{reference index value for start of transmission. Defaults to 0}

\item{mask}{Strength of reduction in transmission rate. Defaults to 0 (i.e. no effect)}

\item{grad}{Strength of decline in transmission once begun. Defaults to 0.5}
}
\description{
Sigmoid function to allow drop in transmission rate
}
\keyword{Decline}
\keyword{Sigmoid}
