% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_DENV_MCMC.R
\name{DENV_mcmc}
\alias{DENV_mcmc}
\title{R Code for SEIR DENV epidemic in a single population}
\usage{
DENV_mcmc(y.vals, time.vals, MCMC.runs)
}
\arguments{
\item{y.vals}{data series}

\item{time.vals}{time series}

\item{MCMC.runs}{number of iterations for MCMC loop}
}
\description{
Single SEIR epidemic
}
