% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ggplottheme_zikafiji.R
\name{theme_zika_fiji}
\alias{theme_zika_fiji}
\title{Background theme for ggplots for Zika paper}
\usage{
theme_zika_fiji(base_size = 11, base_family = "")
}
\description{
Background theme for ggplots for Zika paper
}
